added comments and clean up some comments and removed unused code
