+++ server/src/common/core.ts
@@ -1,8 +1,12 @@
 /*
-* core.ts
-* Copyright (c) Microsoft Corporation.
-* Licensed under the MIT license.
-*/
+ * core.ts
+ * Copyright (c) Microsoft Corporation.
+ * Licensed under the MIT license.
+ *
+ * Various misc code snippets that doesn't have any dependency to other user code.
+ * This is a place one can put code to break circular references between files.
+ * make sure this doesn't have any dependency to other files.
+ */
 
 export const enum Comparison {
     LessThan = -1,
