+++ server/src/tests/checker.test.ts
@@ -1,11 +1,11 @@
 /*
-* checker.test.ts
-* Copyright (c) Microsoft Corporation.
-* Licensed under the MIT license.
-* Author: Eric Traut
-*
-* Unit tests for pyright type checker and type analyzer.
-*/
+ * checker.test.ts
+ * Copyright (c) Microsoft Corporation.
+ * Licensed under the MIT license.
+ * Author: Eric Traut
+ *
+ * Unit tests for pyright type checker and type analyzer.
+ */
 
 import * as assert from 'assert';
