+++ server/src/tests/collectionUtils.test.ts
@@ -1,8 +1,8 @@
 /*
-* collectionUtils.test.ts
-* Copyright (c) Microsoft Corporation.
-* Licensed under the MIT license.
-*/
+ * collectionUtils.test.ts
+ * Copyright (c) Microsoft Corporation.
+ * Licensed under the MIT license.
+ */
 
 import * as assert from 'assert';
 import * as utils from '../common/collectionUtils';
