+++ server/src/tests/debug.test.ts
@@ -1,8 +1,8 @@
 /*
-* debug.test.ts
-* Copyright (c) Microsoft Corporation.
-* Licensed under the MIT license.
-*/
+ * debug.test.ts
+ * Copyright (c) Microsoft Corporation.
+ * Licensed under the MIT license.
+ */
 
 import * as assert from 'assert';
 import * as debug from '../common/debug';
