+++ server/src/tests/harness/fourslash/runner.ts
@@ -1,20 +1,36 @@
 /*
-* runner.ts
-* Copyright (c) Microsoft Corporation.
-* Licensed under the MIT license.
-*/
+ * runner.ts
+ * Copyright (c) Microsoft Corporation.
+ * Licensed under the MIT license.
+ * 
+ * Provide APIs to run fourslash tests from provided fourslash markup contents
+ */
 
-import * as io from "../io"
+import * as ts from "typescript";
 import { combinePaths } from "../../../common/pathUtils";
+import * as io from "../io";
 import { parseTestData } from "./fourSlashParser";
 import { TestState } from "./testState";
-import * as ts from "typescript"
 
+/** 
+ * run given fourslash test file
+ * 
+ * @param basePath this is used as a base path of the virtual file system the test will run upon
+ * @param fileName this is the file path where fourslash test file will be read from
+ */
 export function runFourSlashTest(basePath: string, fileName: string) {
     const content = (io.IO.readFile(fileName)!);
     runFourSlashTestContent(basePath, fileName, content);
 }
 
+/**
+ * run given fourslash markup content
+ * 
+ * @param basePath  this is used as a base path of the virtual file system the test will run upon
+ * @param fileName this will be used as a filename of the given `content` in the virtual file system
+ *                 if fourslash markup `content` doesn't have explicit `@filename` option
+ * @param content  this is fourslash markup string
+ */
 export function runFourSlashTestContent(basePath: string, fileName: string, content: string) {
     // give file paths an absolute path for the virtual file system
     const absoluteBasePath = combinePaths("/", basePath);
