+++ server/src/tests/parser.test.ts
@@ -1,12 +1,12 @@
 /*
-* parser.test.ts
-* Copyright (c) Microsoft Corporation.
-* Licensed under the MIT license.
-* Author: Eric Traut
-*
-* Unit tests for Python parser. These are very basic because
-* the parser gets lots of exercise in the type checker tests.
-*/
+ * parser.test.ts
+ * Copyright (c) Microsoft Corporation.
+ * Licensed under the MIT license.
+ * Author: Eric Traut
+ *
+ * Unit tests for Python parser. These are very basic because
+ * the parser gets lots of exercise in the type checker tests.
+ */
 
 import * as assert from 'assert';
