+++ server/src/tests/stringUtils.test.ts
@@ -1,8 +1,8 @@
 /*
-* stringUtils.test.ts
-* Copyright (c) Microsoft Corporation.
-* Licensed under the MIT license.
-*/
+ * stringUtils.test.ts
+ * Copyright (c) Microsoft Corporation.
+ * Licensed under the MIT license.
+ */
 
 import * as assert from 'assert';
 import * as utils from '../common/stringUtils';
